% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{cms}
\alias{cms}
\title{Means corresponding to lognormal distribution (for generating sources)}
\format{A vector with 10 elements}
\description{
On log scale
}
\details{
\itemize{
		\item Vector of means for 10 sources
		...
}
}
\keyword{datasets}

