% Generated by roxygen2 (4.0.1): do not edit by hand
\name{msefun}
\alias{msefun}
\title{\code{msefun} Get MSE for results}
\usage{
msefun(percinc, etas, rn)
}
\arguments{
\item{percinc}{results for percent increase}

\item{etas}{vector of % increases for health effects}

\item{rn}{rownames}
}
\description{
\code{msefun} Get MSE for results
}

