% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reorderout}
\alias{reorderout}
\title{\code{reorderout} Get outcome}
\usage{
reorderout(out, nr, sources)
}
\arguments{
\item{out}{regression outcomes}

\item{sources}{sources}

\item{rn}{rownames}
}
\description{
\code{reorderout} Get outcome
}

