% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{sds}
\alias{sds}
\title{Standard deviations corresponding to lognormal distribution (for generating sources)}
\format{A vector with 10 elements}
\description{
On log scale
}
\details{
\itemize{
		\item Vector of standard deviations for 10 sources
		...
}
}
\keyword{datasets}

