% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{vec}
\alias{vec}
\title{Positive-part, varimax rotated PCs from Allegheny county for simulating source data}
\format{A matrix with 24 rows and 7 variables}
\description{
Each row is a constituent
}
\details{
\itemize{
		\item PC1. positive-part, varimax rotated PC
    	\item PC2. positive-part, varimax rotated PC
     \item PC3. positive-part, varimax rotated PC
     \item PC4. positive-part, varimax rotated PC
     \item PC5. positive-part, varimax rotated PC
     \item PC6. positive-part, varimax rotated PC
     \item PC7. positive-part, varimax rotated PC
		...
}
}
\keyword{datasets}

